requirements.txt
torch
torchvision
tqdm
accelerate
wandb
numpy
pillow